Create DictionaryLibrary.js and hook it up _Host.cshtml in order to add JS functionality in dynamically changing page url to reflect searched term
Currently the URL of the page is static, and this JS function grants abillity to swap the page url dynamically.  Reason this is useful is that the user can search for a term and the URL will reflect the search, so the URL can be saved or sent somewhere else.