Implement handling words that do not have any kanji associated with them
In Japanese there are certain words that are written without Chinese characters (kanji), and this change makes it so there's a conditional to check if this is the case.

If there are no chinese characters, the main full-word will use the reading property to display it, and empty character instead of reading.