Refresh the design of the search bar for better mobile experience
The new design of the search bar features an icon instead of text saying: "Search" which makes it look a lot slimmer.
Removes border when the bar is selected for nicer appearance.
By using display: flex in the wrapper around the search bar, it makes the mobile experience a lot better as it won't  collapse in on itself, if the user is using the web-app on a small phone.