Convert Search term to lower-case in order to exclude 'advanced' title search
The API attempts to find the most similar results to the search query, and if you search anything with an upper-case first letter, it has a tendency to retrieve unwanted results.
For instance, searching "Cat" would give you Catalyst, instead of expected animal cat, where as "cat" would have given the right word.
This change makes all of the searches automatically lower-case, to avoid getting given names for things or places.